backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
openai==1.42.0
python-dotenv==1.0.1
tenacity==8.5.0
pytest==8.3.2
